// lib/content/types.ts
export type Exercise =
  | MultipleChoiceExercise
  | TranslationExercise
  | FillInTheBlankExercise
  | MatchingPairsExercise
  | ListeningExercise
  | WordBankExercise
  | SpeakingExercise
  | FreeTextExercise
  | FlashcardReviewExercise;

export interface MultipleChoiceExercise {
  type: "multiple-choice";
  text: string;
  choices: string[];
  correctIndex: number;
  randomOrder?: boolean;
  noAudio?: string[];
  srsWords: string | string[];
}

export interface TranslationExercise {
  type: "translation";
  text: string;
  sentence: string;
  answer: string;
  acceptAlso: string[];
  noAudio?: string[];
  srsWords: string | string[];
}

export interface FillInTheBlankExercise {
  type: "fill-in-the-blank";
  sentence: string;
  blank: string;
  noAudio?: string[];
  srsWords: string | string[];
}

export interface MatchingPairsExercise {
  type: "matching-pairs";
  pairs: { left: string; right: string }[];
  randomOrder?: boolean;
  noAudio?: string[];
  srsWords: string | string[];
}

export interface ListeningExercise {
  type: "listening";
  text: string;
  ttsLang: string;
  mode?: "choices" | "word-bank";
  choices?: string[];
  correctIndex?: number;
  noAudio?: string[];
  srsWords: string | string[];
}

export interface WordBankExercise {
  type: "word-bank";
  text: string;
  words: string[];
  answer: string[];
  randomOrder?: boolean;
  noAudio?: string[];
  srsWords: string | string[];
}

export interface SpeakingExercise {
  type: "speaking";
  sentence: string;
  noAudio?: string[];
  srsWords: string | string[];
}

export interface FreeTextExercise {
  type: "free-text";
  text: string;
  afterSubmitPrompt: string;
  noAudio?: string[];
  srsWords?: string | string[];
}

export interface FlashcardReviewExercise {
  type: "flashcard-review";
  front: string;
  back: string;
  noAudio?: string[];
  srsWords: string | string[];
}